+++ src/main/java/net/brnbrd/delightful/compat/Modid.java
@@ -28,6 +28,7 @@ public enum Modid {
 	AS("appleskin"),
 	AT("atmospheric"),
 	ATM("allthemodium"),
+	AUT("autumnity"),
 	BB("buzzier_bees"),
 	BC("brewinandchewin"),
 	BG("berry_good"),